Add Android support to the Cluster Telemetry scripts.
BUG=skia:2489

Review URL: https://codereview.chromium.org/262593008

+++ cluster_telemetry/telemetry_master_scripts/vm_utils.sh
@@ -31,6 +31,21 @@ function is_slave_currently_executing() {
   echo false
 }
 
+function build_chrome_shell_apk {
+  cd $CHROMIUM_BUILD_DIR_BASE
+  echo "{ 'GYP_DEFINES': 'OS=android', }" > chromium.gyp_env
+  cd src
+  . build/android/envsetup.sh
+  GYP_GENERATORS='ninja' ./build/gyp_chromium
+  /b/depot_tools/ninja -C out/Release chrome_shell_apk
+  if [ $? -ne 0 ]
+  then
+    echo "There was an error building chrome_shell_apk $CHROMIUM_COMMIT_HASH + skia $SKIA_COMMIT_HASH"
+    copy_build_log_to_gs
+    exit 1
+  fi
+}
+
 function build_content_shell {
   GYP_GENERATORS='ninja' ./build/gyp_chromium
   /b/depot_tools/ninja -C out/Release content_shell
@@ -66,16 +81,22 @@ function copy_build_to_google_storage {
   chromium_build_dir=$2
   # cd into the Release directory.
   cd $chromium_build_dir/src/out/Release
-  # Move the not needed large directories.
-  mv gen /tmp/
-  mv obj /tmp/
-  # Clean the directory in Google Storage.
-  gsutil rm -R gs://chromium-skia-gm/telemetry/chromium-builds/$dir_name/*
-  # Copy the newly built chrome binary into Google Storage.
-  gsutil cp -r * gs://chromium-skia-gm/telemetry/chromium-builds/$dir_name/
-  # Move the large directories back.
-  mv /tmp/gen .
-  mv /tmp/obj .
+
+  if [ "$TARGET_PLATFORM" == "Android" ]; then
+    # For android builds only need to copy the apks directory to Google Storage.
+    gsutil cp -r apks gs://chromium-skia-gm/telemetry/chromium-builds/$dir_name/
+  else
+    # Move the not needed large directories.
+    mv gen /tmp/
+    mv obj /tmp/
+    # Clean the directory in Google Storage.
+    gsutil rm -R gs://chromium-skia-gm/telemetry/chromium-builds/$dir_name/*
+    # Copy the newly built chrome binary into Google Storage.
+    gsutil cp -r * gs://chromium-skia-gm/telemetry/chromium-builds/$dir_name/
+    # Move the large directories back.
+    mv /tmp/gen .
+    mv /tmp/obj .
+  fi
 }
 
 function apply_patch {
